CHG: Code modified to be able to compile for MDI. BUG: PDelay wasn't calculated correctly as the neighbor rate ratio wasn't used correctly. BUG: PDelay responses used the clock of the time aware system as its clock identity and not the one of the network port. BUG: The path trace TLV of announce messages wasn't always set correctly. BUG: Sync messages weren't processed if the time aware system wasn't grandmaster capable (priority1 == 255). CHG: Time control does a time jump if the offset to the master exceeds 10 ms (formerly, it was 1 ms). CHG: Different time control parameters for MDI. BUG: It could happen that the time offset of x540 wasn't controlled correctly with PTSS. Changed "clock_adjtime" to a syscall to be able to compile the code on systems where this method isn't defined.

+++ timesync_new/common/statemachines/portipc.cpp
@@ -60,7 +60,7 @@ void PortIPC::ProcessState()
         localSystemPhaseOffset = systemTime - deviceTime;
 
         if(m_timeSystemPrevious > 0 && m_timeDevicePrevious > 0 && systemTime - m_timeSystemPrevious > 0)
-            localSystemFrequencyOffset = (deviceTime - m_timeDevicePrevious) / (systemTime - m_timeSystemPrevious);
+            localSystemFrequencyOffset = (float)(deviceTime - m_timeDevicePrevious) / (float)(systemTime - m_timeSystemPrevious);
         else
             localSystemFrequencyOffset = 1.0;
